+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Changed
+
+- Improved redirection of D3D12Core.dll by intercepting the low level function LdrLoadDll instead of LoadLibraryW.
+
 ## 0.3.0 - 2024-07-19
 
 ### Added